move generate-scripts to src/generate
